32 hw12 list groups

+++ OtusHomeWork2026/TelegramBot/ScenariosTasks/AddTaskScenario.cs
@@ -1,12 +1,14 @@
 ﻿using OtusHomeWork2026.Core.DataAccess;
+using OtusHomeWork2026.Core.Dto;
 using OtusHomeWork2026.Core.Entities;
+using OtusHomeWork2026.Core.ScenariosCore;
 using OtusHomeWork2026.Core.Services;
-using OtusHomeWork2026.TelegramBot.Scenarios;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 using Telegram.Bot;
 using Telegram.Bot.Types;
@@ -18,74 +20,117 @@ namespace OtusHomeWork2026.TelegramBot.ScenariosTasks
     {
         IToDoService _toDoService;
         IUserService _userService;
+        IToDoListService _toDoListService;
         ToDoItem _toDoitem;
+        string inputUserData = string.Empty;
+
         string formatDeadLine = "dd.MM.yyyy";
-        public AddTaskScenario(IToDoService toDoService, IUserService userService)
+        public AddTaskScenario(IToDoService toDoService, IUserService userService, IToDoListService toDoListService)
         {
             _toDoService = toDoService;
             _userService = userService;
+            _toDoListService = toDoListService;
         }
         public bool CanHandle(ScenarioType scenario)
         {
             return scenario == ScenarioType.Add;
         }
 
-        public async Task<ScenarioResult> HandleMessageAsync(ITelegramBotClient bot, ScenarioContext context, Message message, CancellationToken ct)
+        public async Task<ScenarioResult> HandleMessageAsync(ITelegramBotClient bot, ScenarioContext context, Update update, CancellationToken ct)
         {
             var scenarioResult = ScenarioResult.Transition;
-            var toDoUser = await _userService.GetUserAsync(message.From.Id, ct);
-            ReplyKeyboardMarkup _replyKeyboard = CreateCanselKeyboard();
+            
+            var toDoUser = await _userService.GetUserAsync(UpdateHandler.GetUserIdFromUpdate(update), ct);
+            ReplyKeyboardMarkup _replyKeyboard = Const.CreateCanselKeyboard();
             ReplyKeyboardMarkup _replyKeyboardDefault = Const.CreateReplyKeyboardMarkup(toDoUser);
-
-            var inputUserData = message.Text;
+            
+            if (update.Message != null)
+                inputUserData = update.Message.Text;
             if (inputUserData == Const.CmCansel)
                 context.CurrentStep = "Cancel";
+
 
             switch (context.CurrentStep)
             {
                 case null:
                     context.Data.Add(toDoUser.TelegramUserId.ToString(), toDoUser);
-                    await bot.SendMessage(message.Chat, "Введите название задачи:", replyMarkup: _replyKeyboard, cancellationToken: ct);
-                    context.CurrentStep = "Name";
+                    await bot.SendMessage(update.Message.Chat, "Введите название задачи:", replyMarkup: _replyKeyboard, cancellationToken: ct);
+                    context.CurrentStep = "ToList";
                     break;
-                case "Name":
+                case "ToList":
                     if (string.IsNullOrEmpty(inputUserData))
                     {
                         await bot.SendMessage(
-                            message.Chat,
+                            update.Message.Chat,
                             $"Нужно добавить описание задачи.",
                             cancellationToken: ct);
                         break;
                     }
                     else
                     {
-                        _toDoitem = await _toDoService.AddAsync(toDoUser, inputUserData, ct);
-                        context.CurrentStep = "Deadline";
-                        await bot.SendMessage(message.Chat, $"Введите срок выполнения {formatDeadLine}:", replyMarkup: _replyKeyboard, cancellationToken: ct);
-                        break;
+
+                        var userLists = await _toDoListService.GetUserLists(toDoUser.UserId, ct);
+                        InlineKeyboardMarkup inlineKeyboard = new InlineKeyboardMarkup(
+                                                                            InlineKeyboardButton.WithCallbackData(
+                                                                                text: "📌 Без списка",
+                                                                                callbackData: "noList"
+                                                                            )
+                                                                      );
+                        if (userLists != null)
+                            foreach (var list in userLists)
+                            {
+                                var toDoListCallbackDto = ToDoListCallbackDto.FromString($"{list.Id}");
+                                inlineKeyboard.AddNewRow(
+                                    new[]
+                                    {
+                                            InlineKeyboardButton.WithCallbackData(text: list.Name, callbackData: toDoListCallbackDto.ToString()),
+                                    });
+                            }
+                        await bot.SendMessage(update.Message.Chat, "Выберите список для добавления", replyMarkup: inlineKeyboard, cancellationToken: ct);
+
+                        context.CurrentStep = "Name";
                     }
-                        
+                    break;
+                case "Name":
+
+                    ToDoList? listData = Guid.TryParse(update.CallbackQuery.Data.Split("|")[0], out var guid)
+                    ? await _toDoListService.Get(guid, ct)
+                    : null;
+                    var name = listData == null ? "📌 Без списка" : listData.Name;
+                    await bot.EditMessageText(
+                            UpdateHandler.GetChatFromUpdate(update),
+                            update.CallbackQuery.Message.Id,
+                            $"Выбран список '{name}'",
+                            cancellationToken: ct
+                        );
+                    _toDoitem = await _toDoService.AddAsync(toDoUser, inputUserData, listData, ct);
+                    context.CurrentStep = "Deadline";
+                    await bot.SendMessage(UpdateHandler.GetChatFromUpdate(update), $"Введите срок выполнения {formatDeadLine}:", replyMarkup: _replyKeyboard, cancellationToken: ct);
+                    break;
+                    
                 case "Deadline":
                     DateTime deadline;
                     DateTime.TryParseExact(inputUserData, formatDeadLine, CultureInfo.InvariantCulture, DateTimeStyles.None, out deadline);
                     if (deadline == DateTime.MinValue)
                     {
-                        await bot.SendMessage(message.Chat, $"Введите срок выполнения {formatDeadLine}:", replyMarkup: _replyKeyboard, cancellationToken: ct);
+                        await bot.SendMessage(update.Message.Chat, $"Введите срок выполнения {formatDeadLine}:", replyMarkup: _replyKeyboard, cancellationToken: ct);
                         break;
                     }
 
                     _toDoitem.DeadLine = deadline;
                     scenarioResult = ScenarioResult.Completed;
-                    await bot.SendMessage(message.Chat, "Задача добавлена.", replyMarkup: _replyKeyboardDefault, cancellationToken: ct);
+                    await bot.SendMessage(update.Message.Chat, "Задача добавлена.", replyMarkup: _replyKeyboardDefault, cancellationToken: ct);
                     break;
                 case "Cancel":
-                    var task = (await _toDoService.FindAsync(toDoUser, _toDoitem.TaskName, ct)).FirstOrDefault();
-                    if (task != null)
-                        await _toDoService.DeleteAsync(task.GuidId, ct);
-
+                    if (_toDoitem != null)
+                    {
+                        var task = (await _toDoService.FindAsync(toDoUser, _toDoitem.TaskName, ct)).FirstOrDefault();
+                        if (task != null)
+                            await _toDoService.DeleteAsync(task.GuidId, ct);
+                    }
                     scenarioResult = ScenarioResult.Completed;
                     context.CurrentStep = "Сценарий завершен.";
-                    await bot.SendMessage(message.Chat, "Операция отменена.", replyMarkup: _replyKeyboardDefault, cancellationToken: ct);
+                    await bot.SendMessage(update.Message.Chat, "Операция отменена.", replyMarkup: _replyKeyboardDefault, cancellationToken: ct);
                     break;
                 default:
                     break;
@@ -93,10 +138,5 @@ namespace OtusHomeWork2026.TelegramBot.ScenariosTasks
             return scenarioResult;
         }
 
-        private ReplyKeyboardMarkup CreateCanselKeyboard ()
-        {
-
-            return new ReplyKeyboardMarkup(new KeyboardButton(Const.CmCansel)) { ResizeKeyboard = true};
-        }
     }
 }
